refactor: Consolidate screen effect logic and update imports
This commit consolidates recurring screen effect logic into a new `ScreenEffect` composable.
This reduces code duplication across various screens.

Additionally, this commit reorganizes imports by moving utility functions like
`getHorizontalAlignment`, `getVerticalArrangement`, `onDoubleTap`, `onPress`,
`PAGE_INDICATOR_HEIGHT`, and `DRAG_HANDLE_SIZE` from `HomeUtil.kt` to a new
`ScreenUtil.kt` file. This improves organization and clarity of utility functions.

The `handleWallpaperScrollEffect` function has been moved from `PageUtil.kt` to
`PagerScreenState.kt` to be closer to its usage context.

Affected files:
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/Screen.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/PagerScreenState.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/application/EblanApplicationInfoGridItem.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/application/PrivateSpaceScreen.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/application/list/ListApplicationScreen.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/application/vertical/DragAndDropEblanApplicationInfos.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/folder/InteractiveFolderGridItem.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/InteractiveGridItem.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/PagerScreen.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/shortcutconfig/ShortcutConfigScreen.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/widget/WidgetScreen.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/GridUtil.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/util/HomeUtil.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/util/PageUtil.kt
- feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/ScreenUtil.kt
