+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/util/HomeUtil.kt
@@ -20,16 +20,10 @@ package com.eblan.launcher.feature.home.util
 import android.content.Context
 import android.content.Intent
 import android.graphics.Rect
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.AnimationVector1D
-import androidx.compose.foundation.gestures.PressGestureScope
-import androidx.compose.ui.unit.dp
 import com.eblan.launcher.domain.model.EblanAction
 import com.eblan.launcher.domain.model.EblanActionType
 import com.eblan.launcher.domain.model.GlobalAction
 import com.eblan.launcher.framework.launcherapps.AndroidLauncherAppsWrapper
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
 
 internal fun handleEblanAction(
     context: Context,
@@ -90,41 +84,5 @@ internal fun handleEblanAction(
     }
 }
 
-internal fun onDoubleTap(
-    context: Context,
-    doubleTap: EblanAction,
-    launcherApps: AndroidLauncherAppsWrapper,
-    scope: CoroutineScope,
-    onOpenAppDrawer: () -> Unit,
-) {
-    if (doubleTap.eblanActionType == EblanActionType.None) return
-
-    scope.launch {
-        handleEblanAction(
-            context = context,
-            eblanAction = doubleTap,
-            launcherApps = launcherApps,
-            onOpenAppDrawer = onOpenAppDrawer,
-        )
-    }
-}
-
-internal suspend fun PressGestureScope.onPress(
-    isVisibleOverlay: Boolean,
-    scale: Animatable<Float, AnimationVector1D>,
-) {
-    if (isVisibleOverlay) return
-
-    awaitRelease()
-
-    if (scale.isRunning) {
-        scale.stop()
-
-        scale.animateTo(1f)
-    }
-}
-
-internal val PAGE_INDICATOR_HEIGHT = 30.dp
-internal val DRAG_HANDLE_SIZE = 30.dp
 internal const val FOLDER_PREVIEW_COLUMNS = 2
 internal const val FOLDER_PREVIEW_ROWS = 2
