+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/util/PageUtil.kt
@@ -17,73 +17,9 @@
  */
 package com.eblan.launcher.feature.home.util
 
-import androidx.compose.foundation.pager.PagerState
-import androidx.compose.runtime.snapshotFlow
-import com.eblan.launcher.framework.wallpapermanager.AndroidWallpaperManagerWrapper
-import kotlinx.coroutines.flow.onStart
-import kotlin.math.absoluteValue
-
 internal fun calculatePage(index: Int, infiniteScroll: Boolean, pageCount: Int): Int = if (infiniteScroll) {
     val offsetIndex = index - (Int.MAX_VALUE / 2)
     offsetIndex - offsetIndex.floorDiv(pageCount) * pageCount
 } else {
     index
-}
-
-internal suspend fun handleWallpaperScrollEffect(
-    horizontalPagerState: PagerState,
-    infiniteScroll: Boolean,
-    pageCount: Int,
-    wallpaperManagerWrapper: AndroidWallpaperManagerWrapper,
-    wallpaperScroll: Boolean,
-    windowToken: android.os.IBinder,
-) {
-    if (!wallpaperScroll) return
-
-    var reverseXOffset: Float
-
-    snapshotFlow { horizontalPagerState.currentPageOffsetFraction }.onStart {
-        wallpaperManagerWrapper.setWallpaperOffsetSteps(
-            xStep = 1f / (pageCount - 1),
-            yStep = 1f,
-        )
-    }.collect { offsetFraction ->
-        val page = calculatePage(
-            index = horizontalPagerState.currentPage,
-            infiniteScroll = infiniteScroll,
-            pageCount = pageCount,
-        )
-
-        val scrollProgress = page + offsetFraction
-
-        if (scrollProgress < 0f) {
-            reverseXOffset = offsetFraction.absoluteValue
-
-            wallpaperManagerWrapper.setWallpaperOffsets(
-                windowToken = windowToken,
-                xOffset = reverseXOffset,
-                yOffset = 0f,
-            )
-        } else if (scrollProgress > pageCount - 1) {
-            reverseXOffset = 1f - offsetFraction
-
-            wallpaperManagerWrapper.setWallpaperOffsets(
-                windowToken = windowToken,
-                xOffset = reverseXOffset,
-                yOffset = 0f,
-            )
-        } else {
-            val xOffset = scrollProgress / (pageCount - 1)
-
-            wallpaperManagerWrapper.setWallpaperOffsets(
-                windowToken = windowToken,
-                xOffset = xOffset,
-                yOffset = 0f,
-            )
-        }
-
-        if (offsetFraction == 0f) {
-            reverseXOffset = offsetFraction
-        }
-    }
 }
